requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
paho-mqtt==1.6.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
